replace WalletId internal 'UUID' with a 'Digest Blake2b_160'
We needed a way to compute wallet ids in a deterministic manner, without taking the risk of exposing
the root private key. Since the mnemonic (and therefore, the root key) is the only real common piece
of information between two wallets, we got to derive _something_ from it in a deterministic manner.

The original idea was to derive a number / seed that would seed a UUID random generator, though, we
can also just stop the process at the seed (no need for the extra step of computing a uuid). Ids are
therefore just a hex-encoded hash of the public key. MD5 would have been sufficient, but going for
Blake2b 160 because... [insert valid reason here]