Move coin selection validation down to the coin selection functions.
After thoughts, it is a bit weird to have the coin selection yield an
invalid coin selection, and later on only check whether this selection
is valid. Instead, we can pretty much ensure that the output produced by
the coin selection is valid IN THE FIRST PLACE.

Still, since the validation is actually backend specific, it requires a
bit of juggling with the type-system. I went for a type family and a
custom error type based on the `t` parameter since this is probably the
easiest and most flexible thing here.