Strengthen the `slotRangeFromTimeRange` property.
When `slotRangeFromTimeRange` doesn't produce a slot range, it should be
because the upper bound of the specified time range is earlier in time
than the start of the blockchain.

This change adds a check to the property test to ensure that this is
indeed the case (and that `slotRangeFromTimeRange` is not returning
`Nothing` inappropriately).