Replace the `development` flag with `werror` flag.
The `cabal check` tool advises that the following compiler flags should
not be enabled by default:

- -Werror
- -O2

Quotes:

Warning: 'ghc-options: -Werror' makes the package easy to break with future
GHC versions because new GHC versions often add new warnings. Alternatively,
if you want to use this, make it conditional based on a Cabal configuration
flag (with 'manual: True' and 'default: False') and enable that flag during
development.

Warning: 'ghc-options: -O2' is rarely needed. Check that it is giving a real
benefit and not just imposing longer compile times on your users.