Specify an upper bound for `base` in the `cabal` file.
The `cabal check` tool advises that Hackage will reject packages that do
not specify an upper bound for the `base` library.

Quote:

Warning: The dependency 'build-depends: base' does not specify an upper bound
on the version number. Each major release of the 'base' package changes the
API in various ways and most packages will need some changes to compile with
it. The recommended practise is to specify an upper bound on the version of
the 'base' package. This ensures your package will continue to build when a
new major version of the 'base' package is released.