Improve the description of cardinality restrictions.
There is no general principle that requires coin selection algorithms to
ensure that any given UTxO entry is used to pay for at most one output.

It is the case that both the Largest-First and Random-Improve algorithms
impose this restriction, but this is done in order to reduce the
complexity of the optimisation problems they attempt to solve.

In principle, we could have an algorithm that relaxes this restriction.
This would make it possible (for example) to pay for multiple smaller outputs
with a single large UTxO entry.

Since there is no general principle that requires a restriction, this
change removes the "Cardinality" subsection from the common "Interface"
section, and creates equivalent "Cardinality" subsections within the
descriptions of the "Largest-First" and "Random-Improve" algorithms.

This change also updates the "Failure Modes" section accordingly.