Revise section on applying a coin selection algorithm to a wallet.
Of course, the concept of a coin selection algorithm can be completely
divorced from the concept of a wallet. So we take a lot of care, for
example, to talk about the /initiall UTxO set rather than the /wallet/ UTxO
set.

But occasionally placing coin selection in the context of a wallet is
helpful, as that is a very common context in which a UTxO appears, and
it gives readers a bit of context for how these algorithms are used.

Response to feedback here:

https://github.com/input-output-hk/cardano-coin-selection/pull/16#discussion_r397885226