correctly manage fee balancing in Migration steps
so far, testing had only been done with 'stable' fee policies where the cost of increasing
and output wouldn't have any influence on the fees. However, Byron comes with some nice fee
calculations where increasing an output may lead to an increase of fee. Therefore, the current
algorithm was in some edge cases looping indefinitely (trying to increase some change, and then
decrease it to pay for fee, and increase it back because of the surplus .. etc ..).

In practice however, Byron nodes do not require perfect balancing (whereas Jörmungandr does). So,
in Byron, the solution is still quite simple: if adding some change increases the fee beyond what
we can actually pay, we simply don't.

We still can't cope with a setup where the node would require strict balancing AND use a fee policy
that is not stable in terms of the output values. In such cases, the current algorithm would
perform pretty bad but... who on earth will really want such setup o.O ?