Make Largest-First pay for outputs collectively.
This change fixes a subtle problem with our implementation of
Largest-First.

== Old implementation (before applying this change)

Rather than considering the total collective value of all outputs when
attempting to find a selection of inputs, the old implementation
considered each output independently, and for each output, attempted to
find a selection of inputs to cover just that output.

This created an unnecessary cardinality restriction, namely that a
single unique input could only be used to pay for at most one unique
output, which meant that any set of n outputs could only be paid for if
there were at least n inputs available. Under this restriction, the
algorithm could fail even if the total value of inputs was sufficient to
cover the required outputs.

== New implementation (after applying this change)

The updated algorithm now considers the total value of all outputs
collectively, rather than considering them individually.

Under this relaxed scheme, it will always be possible to pay for a given
set of outputs of total value v if the total value u of available inputs
satisfies u >= v.