Revise Largest-First unit tests.
Previously, Largest-First would fail to produce a result if the number
of available inputs was lower than the number of requested outputs.

However, under the new relaxed scheme, Largest-First will always succeed
in paying for a given set of outputs of total value v as long as the
total value u of the UTxO satisfies u >= v, regardless of how many
inputs there are.

This change updates several unit tests that previously expected failure
to now expect a successful result.