Revise the Arbitrary instance for CoinSelectionData.
To avoid too many instances of discarded data in properties that make
use of the CoinSelectionData type, incorporate a bias towards generating
sets of inputs that can completely pay for all the outputs.

Additionally, generate coin sizes that have a reasonably high chance of
collision with one another. This makes it more likely that coin
selection algorithms are able to select just enough inputs to pay for
the requested outputs (with zero excess value), which is an extremely
useful edge case.