Correct inconsistent dust inequality checks.
The library currently defines the dust threshold as the maximum size of
a dust coin, rather than the minimum size of a non-dust coin.

It is necessary to define it this way to ensure that zero-valued coins
are always eliminated from adjusted selections. (It is not possible to
specify a dust threshold lower than zero, and therefore not possible to
prevent zero-valued coins from being eliminated.)