Prevent the migration algorithm from creating money.
The migration algorithm should satisfy the following property:

>>> Total value of all inputs >= sum of selection change coins

Before this change, it was possible to cause the algorithm to violate
this property by giving it a single input value lower than the dust
threshold.

This change adds a guard to the `mkCoinSelection` function, preventing
it from creating a change coin if that change coin would be higher than
the total value of all inputs.